refactor code to use baseUrl from env
