+++ packages/location-management/src/configs/env.ts
@@ -1,0 +1,1 @@
+export const API_BASE_URL = process.env.REACT_APP_OPENSRP_API_BASE_URL;
