+++ packages/location-management/src/constants.tsx
@@ -1,5 +1,3 @@
-export const API_BASE_URL = process.env.REACT_APP_OPENSRP_API_BASE_URL;
-
 //
 // Location Unit
 //
